+++ be/src/main/java/com/yumst/be/batch/config/RestaurantBatchConfig.java
@@ -1,7 +1,7 @@
-package com.yumst.be.restaurant.config;
+package com.yumst.be.batch.config;
 
 import com.yumst.be.restaurant.domain.Restaurant;
-import com.yumst.be.restaurant.dto.RestaurantCSVDto;
+import com.yumst.be.batch.dto.RestaurantCSVDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.repository.JobRepository;
